requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
zope.event==5.0
zope.interface==6.1
